* cherry-pick 1f6b933890d873915a5b156077ff53f1a40050e4

* Update example materials

- Update standard_surface_default bindings to match the current definition of the shader.
- Remove standard_surface_velvet bindings that have no visible effect (clearcoat properties that are set without a clearcoat weight).

* Update to OpenImageIO 2

- Update the OiioImageLoader class to the OpenImageIO 2 API.  For simplicity, this changelist removes support for OpenImageIO 1.x, but supporting both API versions is an option for the future.
- Additional updates and cleanup for OiioImageLoader.

* Update to OpenImageIO 2

- Update the OiioImageLoader class to the OpenImageIO 2 API.  For simplicity, this changelist removes support for OpenImageIO 1.x, but supporting both API versions is an option for the future.
- Additional updates and cleanup for OiioImageLoader.

* cherry-pick 9100ac81231d87f7fbf4dc32f7030867e466bc41

* patch cherry-pick

* cherry-pick merge bafdf8c5ea5cc2d73b8a90d00be05d3fdfa56bfb

* Patches for integrations.

+++ source/MaterialXRender/Mesh.h
@@ -15,7 +15,7 @@ namespace MaterialX
 {
 
 /// Geometry index buffer
-using MeshIndexBuffer = vector<unsigned int>;
+using MeshIndexBuffer = vector<uint32_t>;
 /// Float geometry buffer
 using MeshFloatBuffer = vector<float>;
 
@@ -59,9 +59,9 @@ class MeshStream
     }
 
     /// Resize data to an given number of elements
-    void resize(unsigned int elementCount)
+    void resize(size_t elementCount)
     {
-        _data.resize((size_t) elementCount * (size_t) _stride);
+        _data.resize(elementCount * (size_t) _stride);
     }
 
     /// Get stream name
@@ -143,7 +143,7 @@ class MeshPartition
     }
 
     /// Resize data to the given number of indices
-    void resize(unsigned int indexCount)
+    void resize(size_t indexCount)
     {
         _indices.resize(indexCount);
     }
